adicionando funcionalidade de exibir texto do card com sobreposicao
