+++ index.html
@@ -102,13 +102,13 @@
                 </div>
             </div>
         </div>
-        <p class="descript-card">eu vou aparecer depois de clicar no botao do card
-            <button class="back">
-                voltar
-            </button>
+        <p class="descript-card">
+            O não pagamento ou o atraso no pagamento das verbas trabalhistas constitui uma violação dos direitos do trabalhador e pode resultar em ações legais contra o empregador. 
         </p>
-    </div>
-    
+        <button class="back">
+            voltar 
+        </button>
+    </div>   
 	<section id="servicos">
 		<div class="servicos__container">
 			<h2>Serviços</h2>
@@ -162,6 +162,6 @@
         <i class="fa fa-whatsapp"></i>
     </a>
 	<script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script src="./js/script.js"></script>
+    <script src="./js/script.js" type="module"></script>
 </body>
 </html>
